Replace bespoke table formatting logic with the neat `table` NPM module.
This module allows you to turn arrays of strings into neatly formatted ASCII
tables. The new code might be longer, but it no longer contains difficult to
understand logic around exact number of spaces to print etc. We can offload all
of that complexity to `table` and focus on the data that we want to put in
instead.

This is a preliminary step towards adding a new column to the table that
contains the difference from the previous week.